feat: 종 목록 조회 기능 구현
